+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+- Rename Add command to Install
+
 ## [0.1.0] - 2024-09-13
 
-- Initial release
+- Gem setup and basic structure
+- Init command
+- List templates command
+- Compile template command
+- Add template command